+++ demos/example-electron/vite.renderer.config.ts
@@ -29,7 +29,11 @@ export default defineConfig((env) => {
         '@powersync/web > event-iterator',
         '@powersync/web > js-logger',
         '@powersync/web > lodash/throttle',
-        '@powersync/web > can-ndjson-stream'
+        '@powersync/web > can-ndjson-stream',
+        '@powersync/web > buffer',
+        '@powersync/web > rsocket-core',
+        '@powersync/web > rsocket-websocket-client',
+        '@powersync/web > cross-fetch'
       ]
     },
     plugins: [