+++ demos/example-webpack/webpack.config.js
@@ -7,6 +7,12 @@ module.exports = {
     filename: 'index.js',
     path: path.join(__dirname, 'dist')
   },
+  externals: {
+    // BSON includes imports to these, but does not have a hard requirement for them to be present.
+    crypto: 'Crypto',
+    stream: 'Stream',
+    vm: 'VM'
+  },
   devtool: 'source-map',
   mode: 'development',
   resolve: {