+++ demos/react-native-supabase-todolist/app.config.ts
@@ -58,7 +58,17 @@ const config: ExpoConfig = {
       'expo-build-properties',
       {
         ios: {
-          deploymentTarget: '13.4'
+          deploymentTarget: '13.4',
+          // TODO: New architecture is currently not yet supported by @journeyapps/react-native-quick-sqlite
+          newArchEnabled: false
+        },
+        android: {
+          minSdkVersion: 23,
+          compileSdkVersion: 34,
+          targetSdkVersion: 34,
+          buildToolsVersion: '34.0.0',
+          // TODO: New architecture is currently not yet supported by @journeyapps/react-native-quick-sqlite
+          newArchEnabled: false
         }
       }
     ],