+++ docs/docusaurus.config.ts
@@ -19,7 +19,12 @@ const plugins = Object.entries(packageMap).map(([id, config]) => [
     parametersFormat: 'table',
     propertiesFormat: 'table',
     enumMembersFormat: 'table',
+    excludeProtected: true,
+    excludePrivate: true,
     indexFormat: 'table',
+    disableSources: true,
+    expandObjects: true,
+    useCodeBlocks: true,
     typeDeclarationFormat: 'table',
     membersWithOwnFile: ['Class', 'Enum', 'Function', 'Interface'],
     textContentMappings: {