+++ demos/angular-supabase-todolist/CHANGELOG.md
@@ -1,5 +1,55 @@
 # PowerSync Angular Demo
 
+## 0.0.19
+
+### Patch Changes
+
+- @powersync/web@1.0.2
+
+## 0.0.18
+
+### Patch Changes
+
+- @powersync/web@1.0.1
+
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies [32dc7e3]
+- Updated dependencies [e86e61d]
+  - @powersync/web@1.0.0
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [c3f29a1]
+  - @powersync/web@0.8.1
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [7943626]
+- Updated dependencies [48cc01c]
+  - @powersync/web@0.8.0
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [62e43aa]
+- Updated dependencies [6b01811]
+  - @powersync/web@0.7.0
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [f5e42af]
+  - @powersync/web@0.6.1
+
 ## 0.0.12
 
 ### Patch Changes