+++ demos/react-native-supabase-group-chat/CHANGELOG.md
@@ -1,5 +1,58 @@
 # react-native-supabase-group-chat
 
+## 1.1.8
+
+### Patch Changes
+
+- Updated dependencies [1b66145]
+  - @powersync/common@1.11.1
+  - @powersync/react-native@1.7.1
+
+## 1.1.7
+
+### Patch Changes
+
+- Updated dependencies [820a81d]
+- Updated dependencies [820a81d]
+  - @powersync/react-native@1.7.0
+  - @powersync/common@1.11.0
+
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies [e86e61d]
+- Updated dependencies [32dc7e3]
+  - @powersync/react-native@1.6.2
+  - @powersync/common@1.10.0
+
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies [48cc01c]
+  - @powersync/react-native@1.6.1
+  - @powersync/react@1.3.5
+
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [62e43aa]
+- Updated dependencies [6b01811]
+  - @powersync/react-native@1.6.0
+  - @powersync/common@1.9.0
+  - @powersync/react@1.3.4
+
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [f5e42af]
+  - @powersync/react-native@1.5.1
+  - @powersync/common@1.8.1
+  - @powersync/react@1.3.3
+
 ## 1.1.2
 
 ### Patch Changes