+++ demos/vue-supabase-todolist/CHANGELOG.md
@@ -1,5 +1,69 @@
 # vue-supabase-todolist
 
+## 0.0.16
+
+### Patch Changes
+
+- 7ac4911: Updated Vuetify dependency, used lab components are now in Vuetify core library.
+  - @powersync/web@1.0.2
+
+## 0.0.15
+
+### Patch Changes
+
+- @powersync/web@1.0.1
+
+## 0.0.14
+
+### Patch Changes
+
+- 4463851: Moved Vue package to beta.
+- e86e61d: Update PowerSync branding
+- Updated dependencies [4463851]
+- Updated dependencies [32dc7e3]
+- Updated dependencies [e86e61d]
+  - @powersync/vue@0.1.6
+  - @powersync/web@1.0.0
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [c3f29a1]
+  - @powersync/web@0.8.1
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [7943626]
+- Updated dependencies [48cc01c]
+  - @powersync/web@0.8.0
+  - @powersync/vue@0.1.5
+
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [62e43aa]
+- Updated dependencies [6b01811]
+  - @powersync/web@0.7.0
+  - @powersync/vue@0.1.4
+
+## 0.0.10
+
+### Patch Changes
+
+- c3588c0: Updated the vite conf include rules for bson, buffer, rsocket and cross-fetch.
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [f5e42af]
+  - @powersync/vue@0.1.3
+  - @powersync/web@0.6.1
+
 ## 0.0.8
 
 ### Patch Changes