+++ demos/yjs-react-supabase-text-collab/CHANGELOG.md
@@ -1,5 +1,65 @@
 # yjs-react-supabase-text-collab
 
+## 0.1.10
+
+### Patch Changes
+
+- @powersync/web@1.0.2
+
+## 0.1.9
+
+### Patch Changes
+
+- @powersync/web@1.0.1
+
+## 0.1.8
+
+### Patch Changes
+
+- e86e61d: Update PowerSync branding
+- Updated dependencies [32dc7e3]
+- Updated dependencies [e86e61d]
+  - @powersync/web@1.0.0
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [c3f29a1]
+  - @powersync/web@0.8.1
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [7943626]
+- Updated dependencies [48cc01c]
+  - @powersync/web@0.8.0
+  - @powersync/react@1.3.5
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [62e43aa]
+- Updated dependencies [6b01811]
+  - @powersync/web@0.7.0
+  - @powersync/react@1.3.4
+
+## 0.1.4
+
+### Patch Changes
+
+- c3588c0: Updated the vite conf include rules for bson, buffer, rsocket and cross-fetch.
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [f5e42af]
+  - @powersync/react@1.3.3
+  - @powersync/web@0.6.1
+
 ## 0.1.2
 
 ### Patch Changes