+++ packages/kysely-driver/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @powersync/kysely-driver
 
+## 0.2.6
+
+### Patch Changes
+
+- e86e61d: Update PowerSync branding
+- Updated dependencies [32dc7e3]
+  - @powersync/common@1.10.0
+
+## 0.2.5
+
+### Patch Changes
+
+- 48cc01c: Reinclude common package as a dependency
+
+## 0.2.4
+
+### Patch Changes
+
+- 6b01811: Add @powersync/common as peer dependency
+- Updated dependencies [62e43aa]
+  - @powersync/common@1.9.0
+
+## 0.2.3
+
+### Patch Changes
+
+- f5e42af: Introduced base tsconfig. SourceMaps are now excluded in dev and release publishes.
+- Updated dependencies [f5e42af]
+  - @powersync/common@1.8.1
+
 ## 0.2.2
 
 ### Patch Changes