+++ packages/web/CHANGELOG.md
@@ -1,5 +1,67 @@
 # @powersync/web
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [1b66145]
+  - @powersync/common@1.11.1
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [820a81d]
+  - @powersync/common@1.11.0
+
+## 1.0.0
+
+### Major Changes
+
+- 32dc7e3: Bump version for Stable release.
+
+### Patch Changes
+
+- e86e61d: Update PowerSync branding
+- Updated dependencies [32dc7e3]
+  - @powersync/common@1.10.0
+
+## 0.8.1
+
+### Patch Changes
+
+- c3f29a1: Created a fix to resolve the issue where watched queries would not update after syncing upstream changes
+
+## 0.8.0
+
+### Minor Changes
+
+- 7943626: Allow package to be used without web workers
+
+### Patch Changes
+
+- 48cc01c: Reinclude common package as a dependency
+
+## 0.7.0
+
+### Minor Changes
+
+- 62e43aa: Improved import and usage of BSON library.
+
+### Patch Changes
+
+- 6b01811: Add @powersync/common as peer dependency
+- Updated dependencies [62e43aa]
+  - @powersync/common@1.9.0
+
+## 0.6.1
+
+### Patch Changes
+
+- f5e42af: Introduced base tsconfig. SourceMaps are now excluded in dev and release publishes.
+- Updated dependencies [f5e42af]
+  - @powersync/common@1.8.1
+
 ## 0.6.0
 
 ### Minor Changes